Add 21.02.1 artifacts, working custom-build.sh + Dockerfile.
This sets up the main plumbing to simply generate the necessary kernel
modules using a OpenWrt-buildfleet compatible config.

It should be pretty damn automated at this point, and easy for anyone to
run themselves.

What remains to be completed is some proper instructions in the README
and a solid `install.sh` script to automate the downloading of these
files, checksum check, and install/downgrade on existing OpenWrt
installs.